Issue #11163: Enforced filesize TypeNoJavadoc1
