+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/missingjavadoctype/InputMissingJavadocTypeNoJavadoc1Two.java
@@ -10,54 +10,7 @@ tokens = (default)INTERFACE_DEF, CLASS_DEF, ENUM_DEF, ANNOTATION_DEF, RECORD_DEF
 
 package com.puppycrawl.tools.checkstyle.checks.javadoc.missingjavadoctype;
 
-public class InputMissingJavadocTypeNoJavadoc1 // violation
-{
-    public int i1;
-    protected int i2;
-    int i3;
-    private int i4;
-
-    public void foo1() {}
-    protected void foo2() {}
-    void foo3() {}
-    private void foo4() {}
-
-    protected class ProtectedInner { // violation
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-
-    class PackageInner { // violation
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-
-    private class PrivateInner { // violation
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-}
+public class InputMissingJavadocTypeNoJavadoc1Two {} // violation
 
 class PackageClass1 { // violation
     public int i1;
